requirements.txt
Flask
Celery
redis
psycopg2-binary